+++ one/alf/io.py
@@ -361,7 +361,7 @@ def _ls(alfpath, object=None, **kwargs) -> (list, tuple):
 
     # raise error if no files found
     if not files_alf:
-        err_str = f'object "{object}" ' if object else 'ALF files'
+        err_str = f'object "{object}"' if object else 'ALF files'
         raise ALFObjectNotFound(f'No {err_str} found in {alfpath}')
 
     return [alfpath.joinpath(f) for f in files_alf], attributes
@@ -501,7 +501,7 @@ def load_object(alfpath, object=None, short_keys=False, **kwargs):
     # Take attribute and timescale from parts list
     attributes = [p[2] if not p[3] else '_'.join(p[2:4]) for p in parts]
     if not short_keys:  # Include extra parts in the keys
-        attributes = [attr + ('.' + p[4] if p[4] else '') for attr, p in zip(attributes, parts)]
+        attributes = ['.'.join(filter(None, (attr, p[4]))) for attr, p in zip(attributes, parts)]
         # TODO List duplicates; raise ALFError
     assert len(set(attributes)) == len(attributes), (
         f'multiple object {object} with the same attribute in {alfpath}, restrict parts/namespace')
@@ -525,8 +525,25 @@ def load_object(alfpath, object=None, short_keys=False, **kwargs):
             # if there is other stuff in the dictionary, save it, otherwise disregard
             if meta:
                 out[att + 'metadata'] = meta
-    if 'table' in out.keys():  # Merge 'table' dataframe into bunch
-        out.update(AlfBunch.from_df(out.pop('table')))
+    # Merge 'table' dataframe into bunch
+    table_key = next(filter(re.compile(r'^table([_.]|$)').match, out), None)  # py 3.8
+    if table_key:
+        table = out.pop(table_key)
+
+        def rename_columns(field):
+            """
+            For each field name in the DataFrame, return a new one that includes any timescale or
+            extra ALF parts found in table_key.
+            For example...
+                with table_key = table_clock, field1 -> field1_clock;
+                with table_key = table_clock.extra, field1_0 -> field1_clock.extra_0;
+                with table_key = table, field1 -> field1
+            """
+            return (field[:-2] + table_key[5:] + field[-2:]
+                    if re.match(r'.+?_[01]$', field)
+                    else field + table_key[5:])
+        table.rename(columns=rename_columns, inplace=True)
+        out.update(AlfBunch.from_df(table))
     status = out.check_dimensions
     timeseries = [k for k in out.keys() if 'timestamps' in k]
     if any(timeseries) and len(out.keys()) > len(timeseries) and status == 0: